backend/requirements.txt
fastapi==0.111.0
PyJWT==2.8.0
alembic==1.13.1
SQLAlchemy==2.0.30
psycopg2==2.9.9
sqlmodel==0.0.18
black==24.4.2
pytest==8.2.2
pika==1.3.2
